+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 - Support extended syntax for configurations of Member Tags and Member Events, to allow multidimensional arrays to be unwrapped (PR #53)
 - Added configuration validation via JSON Schema (PR #54)
 - Support extended configuration for Merge Fields, so that fields can be set to continuously sync even when not changed (PR #55)
-- Dependencies bump (#56)
+- Dependencies bump (PR #56)
+- Support updating mailchimp subscription status via merge field trigger (PR #57)
 
 ## Version 0.4.0